Phase 7: Extract shared respx shim helper
